app/requirements.txt
Flask
pycryptodome
mysql-connector-python
Flask-JWT-Extended
bcrypt